drafts/2020/isochrone/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=19.3.0=py_0
backcall=0.1.0=py38_0
blas=1.0=mkl
bleach=3.1.0=py_0
branca=0.4.0=py_0
bzip2=1.0.8=he774522_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py38_0
cffi=1.14.0=py38ha419a9e_0
cfitsio=3.470=he774522_2
chardet=3.0.4=py38_1003
click=7.0=py_0
click-plugins=1.1.1=py_0
cligj=0.5.0=py38_0
colorama=0.4.3=py_0
cryptography=2.8=py38hb32ad35_1
curl=7.67.0=h2a8f88b_0
cycler=0.10.0=py_2
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
descartes=1.1.0=py_4
entrypoints=0.3=py38_0
expat=2.2.5=he025d50_0
fiona=1.8.4=py38h22081e2_0
folium=0.10.1=py_0
freetype=2.10.0=h563cfd7_1
freexl=1.0.5=hfa6e2cd_0
gdal=2.3.3=py38hdf43c64_0
geographiclib=1.50=py_0
geopandas=0.6.1=py_0
geopy=1.21.0=py_0
geos=3.7.1=h33f27b4_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.9=py_1
importlib_metadata=1.5.0=py38_0
intel-openmp=2020.0=166
ipykernel=5.1.4=py38h39e3cac_0
ipython=7.12.0=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_0
jedi=0.16.0=py38_0
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py38_0
jupyter=1.0.0=py38_7
jupyter_client=5.3.4=py38_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py38_0
kealib=1.4.7=h07cbb95_6
kiwisolver=1.1.0=py38he980bc4_0
krb5=1.16.4=hc04afaa_0
libblas=3.8.0=15_mkl
libboost=1.67.0=hd9e427e_4
libcblas=3.8.0=15_mkl
libcurl=7.67.0=h2a8f88b_0
libgdal=2.3.3=h10f50ba_0
libiconv=1.15=h1df5818_7
libkml=1.3.0=he5f2a48_4
liblapack=3.8.0=15_mkl
libnetcdf=4.6.1=h411e497_2
libpng=1.6.37=h2a8f88b_0
libpq=11.2=h3235a2c_0
libsodium=1.0.16=h9d3ae62_0
libspatialindex=1.9.3=h33f27b4_0
libspatialite=4.3.0a=hc36aec2_19
libssh2=1.9.0=h7a1dbc1_1
libtiff=4.1.0=h56a325e_0
libxml2=2.9.9=h464c3ec_0
lz4-c=1.8.1.2=h2fa13f4_0
m2w64-expat=2.1.1=2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gettext=0.19.7=2
m2w64-gmp=6.1.0=2
m2w64-libiconv=1.14=6
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-xz=5.2.2=2
markupsafe=1.1.1=py38hfa6e2cd_0
matplotlib-base=3.2.0=py38h2981e6d_1
mistune=0.8.4=py38he774522_1000
mkl=2020.0=166
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.0.15=py38h14836fe_0
mkl_random=1.1.0=py38hf9181ef_0
msys2-conda-epoch=20160418=1
munch=2.5.0=py_0
nbconvert=5.6.1=py38_0
nbformat=5.0.4=py_0
networkx=2.4=py_0
notebook=6.0.3=py38_0
numpy=1.18.1=py38h93ca92e_0
numpy-base=1.18.1=py38hc3f5095_1
openjpeg=2.3.0=h5ec785f_1
openssl=1.1.1d=he774522_4
osmnx=0.11.4=py_1
pandas=1.0.1=py38h47e9c7a_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py38_1
parso=0.6.1=py_0
pcre=8.43=ha925a31_0
pickleshare=0.7.5=py38_1000
pip=20.0.2=py38_1
postgresql=11.2=h3235a2c_0
proj4=5.2.0=ha925a31_1
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
pycparser=2.20=py_0
pygments=2.5.2=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.6=py_0
pyproj=1.9.6=py38h6782396_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.15.7=py38he774522_0
pysocks=1.7.1=py38_0
python=3.8.1=h5fd99cc_8_cpython
python-dateutil=2.8.1=py_0
pytz=2019.3=py_0
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38_0
pyzmq=18.1.1=py38ha925a31_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.6.0=py38_1
requests=2.23.0=py38_0
rtree=0.9.3=py38h21ff451_0
scikit-learn=0.22.2.post1=py38h7208079_0
scipy=1.3.2=py38h582fac2_0
send2trash=1.5.0=py38_0
setuptools=46.0.0=py38_0
shapely=1.6.4=py38h222a598_0
sip=4.19.13=py38ha925a31_0
six=1.14.0=py38_0
sqlite=3.31.1=he774522_0
tbb=2018.0.5=he980bc4_0
terminado=0.8.3=py38_0
testpath=0.4.4=py_0
tiledb=1.6.3=h7b000aa_0
tk=8.6.8=hfa6e2cd_0
tornado=6.0.4=py38hfa6e2cd_0
traitlets=4.3.3=py38_0
urllib3=1.25.7=py38_0
vc=14.1=h0510ff6_4
vincent=0.4.4=py_1
vs2015_runtime=14.16.27012=hf0eaf9b_1
wcwidth=0.1.8=py_0
webencodings=0.5.1=py38_1
wheel=0.34.2=py38_0
widgetsnbextension=3.5.1=py38_0
win_inet_pton=1.1.0=py38_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
xerces-c=3.2.2=ha925a31_0
xz=5.2.4=h2fa13f4_4
zeromq=4.3.1=h33f27b4_3
zipp=2.2.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
